Add JumpCloud Google Workspace IdP setup guide (#26)
* Add JumpCloud Google Workspace IdP setup guide and update dependencies

Introduced a comprehensive guide for configuring Google Workspace as an Identity Provider (IdP) in JumpCloud. Added related documentation, images, and updated dependencies, including upgrading `@iconify/react` from `5.2.1` to `6.0.0`. Incremented the project version to `0.0.4` to reflect these changes.

* Fix incorrect syntax in image element in JumpCloud guide

Removed an extraneous '=' character in the image element. This fixes a syntax issue in the "Use Google Workspace as an IdP" documentation to ensure proper rendering.